Prefer defconfigs from BR2_EXTERNAL/configs
Buildroot configurations stored in BR2_EXTERNAL/configs
cannot be used when there is a name collision with
those included in the upstream.

This is leads into problems when there is a custom
defconfig for a board in the BR2_EXTERNAL/configs
and somebody commits a defconfig of the same name
into the upstream. This behaviour leads to unexpected
results without a reason (you end with unwanted kernel,
unwanted device-tree and several other unwanted
settings which is quite confusing).

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>