dpdk: new package
This patch introduces support of the DPDK library (www.dpdk.org) into
Buildroot. It compiles and installs DPDK libraries on the target and
staging and allows to compiler other applications depending on the DPDK
library. It can also install some basic tools the DPDK provides
(testpmd, python scripts, test suite).

The patch assumes DPDK 2.2 which is not out at the moment (scheduled
to the end of November). However, this version will contain support for
the ARM architecture that is currently being reviewed.

Testing of DPDK on ARM is possible by cloning

 https://github.com/RehiveTech/dpdk arm-support-v5

as the package allows customization of the download process.

The ARM ports can be tested by

 qemu_aarch64_virt_defconfig
 qemu_arm_vexpress_defconfig

I did not test, how it works with x86 as I do not use Buildroot
for this.

The included hashes are calculated locally by downloading the tar.gz
archives by hand.

There are unfortunately some pitfalls:

* it may require to enable PCI, MSIX, UIO in the Linux Kernel
  (some defconfigs does not include as default and it is platform
  dependent as ARMv7 almost does not use PCI)

* when building PCAP PMD driver, the libpcap is required (however,
  this dependency is difficult to express in Buildroot as this depends
  on the DPDK configuration file - similar to the issue above)

* some tools the DPDK provides depend on Python(2) so the user has
  to enable it to install those

I assume, the user knows what he is doing in those cases and configures
the Buildroot in the right way. Or any ideas? Yann?

Signed-off-by: Jan Viktorin <viktorin@rehivetech.com>